+++ cpp/fory/serialization/collection_serializer.h
@@ -549,9 +549,10 @@ struct Serializer<
     ctx.buffer().Grow(static_cast<uint32_t>(max_size));
     uint32_t writer_index = ctx.buffer().writer_index();
     writer_index += ctx.buffer().PutVarUint32(
-          writer_index, static_cast<uint32_t>(total_bytes));
+        writer_index, static_cast<uint32_t>(total_bytes));
     if (total_bytes > 0) {
-      ctx.buffer().UnsafePut(writer_index, vec.data(), static_cast<uint32_t>(total_bytes));
+      ctx.buffer().UnsafePut(writer_index, vec.data(),
+                             static_cast<uint32_t>(total_bytes));
     }
     ctx.buffer().WriterIndex(writer_index + static_cast<uint32_t>(total_bytes));
     return Result<void, Error>();
@@ -845,10 +846,11 @@ template <typename Alloc> struct Serializer<std::vector<bool, Alloc>> {
     constexpr size_t max_size = 8 + vec.size();
     ctx.buffer().Grow(static_cast<uint32_t>(max_size));
     uint32_t writer_index = ctx.buffer().writer_index();
-    writer_index += ctx.buffer().PutVarUint32(writer_index, static_cast<uint32_t>(vec.size()));
+    writer_index += ctx.buffer().PutVarUint32(
+        writer_index, static_cast<uint32_t>(vec.size()));
     for (size_t i = 0; i < vec.size(); ++i) {
       writer_index += i;
-      ctx.buffer().UnsafePutByte(writer_index, vec[i] ? 1 : 0);
+      ctx.buffer().UnsafePutByte(writer_index, static_cast<uint8_t>(vec[i] ? 1 : 0));
     }
     ctx.buffer().WriterIndex(writer_index);
     return Result<void, Error>();
